NetPacket abstraction
