+++ src/Client/Client.cs
@@ -15,7 +15,7 @@ namespace Client
         {
         }
 
-        public override void HandleMessage(NetPacket packet)
+        public override void HandleMessage(NetPacketBase packet)
         {
             Console.WriteLine("Incoming message");
